Package trained denoiser as a Frontier Developer Portal processing unit
Local proof of concept only, not deployed to the live platform:
- frontier-app/: Docker container following Frontier's documented
  mount-path convention, Data Selection Rules grammar, and UI
  Configuration JSON Schema
- src/inference.py: single-series DICOM load/preprocess/inference/write,
  reusing the existing trained model rather than retraining
- Verified end-to-end via real docker build/run against real DICOM data
- Fixes a pre-existing float32/float64 bug in src/dicom_loader.py found
  while building this

+++ src/dicom_loader.py
@@ -32,8 +32,12 @@ def load_dicom_series_as_hu(series_dir: str) -> np.ndarray:
 
     slices = []
     for ds in datasets:
+        # pydicom's RescaleSlope/Intercept are DSfloat (a float subclass);
+        # multiplying by a plain Python float upcasts float32 -> float64 per
+        # numpy's type promotion rules, so the explicit cast at the end is
+        # load-bearing, not redundant with the one above.
         hu = ds.pixel_array.astype(np.float32) * ds.RescaleSlope + ds.RescaleIntercept
-        slices.append(hu)
+        slices.append(hu.astype(np.float32))
 
     return np.stack(slices, axis=0)
